+++ src/hooks/api/daily/useEditDailyAnswer.ts
@@ -1,0 +1,22 @@
+'use client';
+
+import { useMutation } from '@tanstack/react-query';
+
+import { patchDailyAnswer } from '@/api/dailyApis';
+import { useModalStore } from '@/stores/useModalStore';
+
+// 답변 수정
+export const usePatchDailyAnswerMutation = () => {
+  const { open } = useModalStore();
+
+  return useMutation({
+    mutationFn: patchDailyAnswer,
+    onSuccess: () => {
+      open('dailyAnswerEdit', undefined);
+    },
+    onError: () => {
+      // FIXME: 모달로 변경
+      alert('답변 수정에 실패했습니다. 잠시 후 다시 시도해주세요.');
+    },
+  });
+};
